+++ packages/fiber/src/core/events.ts
@@ -1,7 +1,7 @@
 import * as THREE from 'three'
-import { UseStore } from 'zustand'
-import { Instance } from './renderer'
-import { RootState } from './store'
+import type { UseStore } from 'zustand'
+import type { Instance } from './renderer'
+import type { RootState } from './store'
 
 export interface Intersection extends THREE.Intersection {
   eventObject: THREE.Object3D
@@ -60,6 +60,36 @@ export interface EventManager<TTarget> {
 
 export function createEvents(store: UseStore<RootState>) {
   const hovered = new Map<string, DomEvent>()
+  const temp = new THREE.Vector3()
+
+  /** Sets up defaultRaycaster */
+  function prepareRay(event: DomEvent) {
+    const state = store.getState()
+    const { raycaster, mouse, camera, size } = state
+    // https://github.com/pmndrs/react-three-fiber/pull/782
+    // Events trigger outside of canvas when moved
+    const { offsetX, offsetY } = raycaster.computeOffsets?.(event, state) ?? event
+    const { width, height } = size
+    mouse.set((offsetX / width) * 2 - 1, -(offsetY / height) * 2 + 1)
+    raycaster.setFromCamera(mouse, camera)
+  }
+
+  /** Calculates delta */
+  function calculateDistance(event: DomEvent) {
+    const { internal } = store.getState()
+    const dx = event.offsetX - internal.initialClick[0]
+    const dy = event.offsetY - internal.initialClick[1]
+    return Math.round(Math.sqrt(dx * dx + dy * dy))
+  }
+
+  /** Returns true if an instance has a valid pointer-event registered, this excludes scroll, clicks etc */
+  function filterPointerEvents(objects: THREE.Object3D[]) {
+    return objects.filter((obj) =>
+      ['Move', 'Over', 'Enter', 'Out', 'Leave'].some(
+        (name) => ((obj as unknown) as Instance).__r3f.handlers?.[('onPointer' + name) as keyof EventHandlers],
+      ),
+    )
+  }
 
   function makeId(event: Intersection) {
     return (event.eventObject || event.object).uuid + '/' + event.index
@@ -113,5 +143,184 @@ export function createEvents(store: UseStore<RootState>) {
     return intersections
   }
 
-  return { hovered, makeId, intersect, patchIntersects }
+  /**  Handles intersections by forwarding them to handlers */
+  function handleIntersects(intersections: Intersection[], event: DomEvent, callback: (event: DomEvent) => void) {
+    const { raycaster, mouse, camera, internal } = store.getState()
+    // If anything has been found, forward it to the event listeners
+    if (intersections.length) {
+      const unprojectedPoint = temp.set(mouse.x, mouse.y, 0).unproject(camera)
+      const delta = event.type === 'click' ? calculateDistance(event) : 0
+      const releasePointerCapture = (id: number) => (event.target as Element).releasePointerCapture(id)
+      const localState = { stopped: false, captured: false }
+
+      for (const hit of intersections) {
+        const setPointerCapture = (id: number) => {
+          // If the hit is going to be captured flag that we're in captured state
+          if (!localState.captured) {
+            localState.captured = true
+            // The captured hit array is reset to collect hits
+            internal.captured = []
+          }
+          // Push hits to the array
+          if (internal.captured)
+            internal.captured.push(hit)
+            // Call the original event now
+          ;(event.target as Element).setPointerCapture(id)
+        }
+
+        // Add native event props
+        let extractEventProps: any = {}
+        for (let prop in Object.getPrototypeOf(event)) {
+          extractEventProps[prop] = event[prop as keyof DomEvent]
+        }
+
+        let raycastEvent: any = {
+          ...hit,
+          ...extractEventProps,
+          intersections,
+          stopped: localState.stopped,
+          delta,
+          unprojectedPoint,
+          ray: raycaster.ray,
+          camera: camera,
+          // Hijack stopPropagation, which just sets a flag
+          stopPropagation: () => {
+            // https://github.com/pmndrs/react-three-fiber/issues/596
+            // Events are not allowed to stop propagation if the pointer has been captured
+            const cap = internal.captured
+            if (!cap || cap.find((h) => h.eventObject.id === hit.eventObject.id)) {
+              raycastEvent.stopped = localState.stopped = true
+
+              // Propagation is stopped, remove all other hover records
+              // An event handler is only allowed to flush other handlers if it is hovered itself
+              if (hovered.size && Array.from(hovered.values()).find((i) => i.eventObject === hit.eventObject)) {
+                // Objects cannot flush out higher up objects that have already caught the event
+                const higher = intersections.slice(0, intersections.indexOf(hit))
+                cancelPointer([...higher, hit])
+              }
+            }
+          },
+          target: { ...event.target, setPointerCapture, releasePointerCapture },
+          currentTarget: { ...event.currentTarget, setPointerCapture, releasePointerCapture },
+          sourceEvent: event,
+        }
+
+        // Call subscribers
+        callback(raycastEvent as DomEvent)
+        // Event bubbling may be interrupted by stopPropagation
+        if (localState.stopped === true) break
+      }
+    }
+    return intersections
+  }
+
+  function cancelPointer(hits: Intersection[]) {
+    Array.from(hovered.values()).forEach((hoveredObj) => {
+      // When no objects were hit or the the hovered object wasn't found underneath the cursor
+      // we call onPointerOut and delete the object from the hovered-elements map
+      if (!hits.length || !hits.find((hit) => hit.object === hoveredObj.object && hit.index === hoveredObj.index)) {
+        const eventObject = hoveredObj.eventObject
+        const handlers = ((eventObject as unknown) as Instance).__r3f.handlers
+        hovered.delete(makeId(hoveredObj))
+        if (handlers) {
+          // Clear out intersects, they are outdated by now
+          const data = { ...hoveredObj, intersections: hits || [] }
+          handlers.onPointerOut?.(data as ThreeEvent<PointerEvent>)
+          handlers.onPointerLeave?.(data as ThreeEvent<PointerEvent>)
+        }
+      }
+    })
+  }
+
+  const handlePointer = (name: string) => {
+    // Deal with cancelation
+    switch (name) {
+      case 'onPointerLeave':
+      case 'onPointerCancel':
+        return () => cancelPointer([])
+      case 'onLostPointerCapture':
+        return () => ((store.getState().internal.captured = undefined), cancelPointer([]))
+    }
+
+    // Any other pointer goes here ...
+    return (event: DomEvent) => {
+      const { onPointerMissed, internal } = store.getState()
+
+      prepareRay(event)
+
+      // Get fresh intersects
+      const isPointerMove = name === 'onPointerMove'
+      const filter = isPointerMove ? filterPointerEvents : undefined
+      const hits = patchIntersects(intersect(filter), event)
+
+      // Take care of unhover
+      if (isPointerMove) cancelPointer(hits)
+
+      handleIntersects(hits, event, (data: DomEvent) => {
+        const eventObject = data.eventObject
+        const handlers = ((eventObject as unknown) as Instance).__r3f.handlers
+        // Check presence of handlers
+        if (!handlers) return
+
+        if (isPointerMove) {
+          // Move event ...
+          if (handlers.onPointerOver || handlers.onPointerEnter || handlers.onPointerOut || handlers.onPointerLeave) {
+            // When enter or out is present take care of hover-state
+            const id = makeId(data)
+            const hoveredItem = hovered.get(id)
+            if (!hoveredItem) {
+              // If the object wasn't previously hovered, book it and call its handler
+              hovered.set(id, data)
+              handlers.onPointerOver?.(data as ThreeEvent<PointerEvent>)
+              handlers.onPointerEnter?.(data as ThreeEvent<PointerEvent>)
+            } else if (hoveredItem.stopped) {
+              // If the object was previously hovered and stopped, we shouldn't allow other items to proceed
+              data.stopPropagation()
+            }
+          }
+          // Call mouse move
+          handlers.onPointerMove?.(data as ThreeEvent<PointerEvent>)
+        } else {
+          // All other events ...
+          const handler = handlers?.[name as keyof EventHandlers] as (event: ThreeEvent<PointerEvent>) => void
+          if (handler) {
+            // Forward all events back to their respective handlers with the exception of click events,
+            // which must use the initial target
+            if (
+              (name !== 'onClick' && name !== 'onContextMenu' && name !== 'onDoubleClick') ||
+              internal.initialHits.includes(eventObject)
+            ) {
+              handler(data as ThreeEvent<PointerEvent>)
+              pointerMissed(
+                event,
+                internal.interaction.filter((object) => object !== eventObject),
+              )
+            }
+          }
+        }
+      })
+
+      // Save initial coordinates on pointer-down
+      if (name === 'onPointerDown') {
+        internal.initialClick = [event.offsetX, event.offsetY]
+        internal.initialHits = hits.map((hit) => hit.eventObject)
+      }
+
+      // If a click yields no results, pass it back to the user as a miss
+      if ((name === 'onClick' || name === 'onContextMenu' || name === 'onDoubleClick') && !hits.length) {
+        if (calculateDistance(event) <= 2) {
+          pointerMissed(event, internal.interaction)
+          if (onPointerMissed) onPointerMissed()
+        }
+      }
+    }
+  }
+
+  function pointerMissed(event: MouseEvent, objects: THREE.Object3D[]) {
+    objects.forEach((object: THREE.Object3D) =>
+      ((object as unknown) as Instance).__r3f.handlers?.onPointerMissed?.(event as ThreeEvent<PointerEvent>),
+    )
+  }
+
+  return { handlePointer }
 }
